added basic drag and drop file functionality
Just a proof of concept to beta test with my classes. If this works okay, should really refactor the keyboard_shortcuts.js file to allow keyboard shortcuts and drag and drop to use the same underlying loading code.

There is also a known bug in this: if you drag and drop to the bottom of the screen, the event doesn't trigger. I *think* this must be something to do with the height in CSS. Haven't explored it enough to know for sure.